safari-600.7-branch/LayoutTests/fast/regex/overflow-expected.txt

This test checks expressions with alternative lengths of appox. 2^31.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regexp1.exec('') is null
PASS regexp2.exec('') is null
PASS regexp3.exec(s3) is null
PASS function f() { /[^a$]{4294967295}/ } threw exception SyntaxError: Invalid regular expression: number too large in {} quantifier.
PASS successfullyParsed is true

TEST COMPLETE
